+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## v1.4.0 (2017-05-20)
+* updated Debian Jessie to version 8.8
+* updated VirtualBox Guest additions to version 5.1.22
+* updated Lets Encrypt Certbot to version 0.14.1
+* updated PostgreSQL to version 9.6.3
+* updated Redis to version 3.2.9
+* updated nginx to version 1.13.0
+* updated OpenSSL to version 1.1.0e
+* updated nginx-rtmp-module to version 1.1.11
+* changed from "dotdeb" to "sury" repository for PHP and PHP-FPM
+* updated PHP and PHP-FPM to version 7.1.5
+* updated PhpRedis to version 3.1.2
+* updated Xdebug to version 2.5.1
+* updated Composer to version 1.4.2
+* updated NodeJS to version 6.10.3
+* updated NPM to version 4.6.1
+* updated Newman to version 3.6.0
+* added Yarn in version 0.24.5
+* updated FFmpeg to version 2.8.11
+* updated UnrealIRCd to version 4.0.12
+* updated Anope to version 2.0.5
+* fixed composer alias behavior in composer.bashrc
+
 ## v1.3.1 (2016-12-22)
 * added apt packages clean
 * updated VirtualBox Guest additions to version 5.1.12